+++ test/services_test.dart
@@ -52,7 +52,7 @@ main() {
     test('can create data', () async {
       String postData = god.serialize({'text': 'Hello, world!'});
       var response = await client.post("$url/todos",
-          headers: headers.cast<String, String>(), body: postData);
+          headers: headers as Map<String, String>, body: postData);
       expect(response.statusCode, 201);
       var json = god.deserialize(response.body);
       print(json);
@@ -62,7 +62,7 @@ main() {
     test('can fetch data', () async {
       String postData = god.serialize({'text': 'Hello, world!'});
       await client.post("$url/todos",
-          headers: headers.cast<String, String>(), body: postData);
+          headers: headers as Map<String, String>, body: postData);
       var response = await client.get("$url/todos/0");
       expect(response.statusCode, 200);
       var json = god.deserialize(response.body);
@@ -73,10 +73,10 @@ main() {
     test('can modify data', () async {
       String postData = god.serialize({'text': 'Hello, world!'});
       await client.post("$url/todos",
-          headers: headers.cast<String, String>(), body: postData);
+          headers: headers as Map<String, String>, body: postData);
       postData = god.serialize({'text': 'modified'});
       var response = await client.patch("$url/todos/0",
-          headers: headers.cast<String, String>(), body: postData);
+          headers: headers as Map<String, String>, body: postData);
       expect(response.statusCode, 200);
       var json = god.deserialize(response.body);
       print(json);
@@ -86,10 +86,10 @@ main() {
     test('can overwrite data', () async {
       String postData = god.serialize({'text': 'Hello, world!'});
       await client.post("$url/todos",
-          headers: headers.cast<String, String>(), body: postData);
+          headers: headers as Map<String, String>, body: postData);
       postData = god.serialize({'over': 'write'});
       var response = await client.post("$url/todos/0",
-          headers: headers.cast<String, String>(), body: postData);
+          headers: headers as Map<String, String>, body: postData);
       expect(response.statusCode, 200);
       var json = god.deserialize(response.body);
       print(json);
@@ -101,7 +101,7 @@ main() {
       String postData = god.serialize({'text': 'Hello, world!'});
       var created = await client
           .post("$url/todos",
-              headers: headers.cast<String, String>(), body: postData)
+              headers: headers as Map<String, String>, body: postData)
           .then((r) => json.decode(r.body));
       var response = await client.delete("$url/todos/${created['id']}");
       expect(response.statusCode, 200);